requirements/local.txt
-r base.txt
-e .
-e ../../app/base
-e ../../app/login
gitpython
hgapi
ipdb
IPy
pytest-django
pytest-flakes
pytest-pep8
walkdir
